requirements/base.txt

# Database Adapter
MySQL-python


# Django stuff
Django

# for bcrypt passwords
bcrypt

# Templates
django_compressor

# Celery: Message queue
celery
django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Logging
logutils

# Mixins
django-braces
django-model-utils
bpython

# Admin
django-debug-toolbar

# Emailsending
sendgrid-django


# Awesome-Admin
django-jet

# Redactor
django-wysiwyg-redactor

#Bootstrap
django-bootstrap3
